lightprof/reasoning.py
from typing import Any, Dict, List, Tuple

import google.generativeai as genai
import torch  # Keep torch import in case future versions handle embeddings directly

from lightprof.adapter import Subgraph  # Import Subgraph type alias

# Define a type alias for the LLM response
LLMResponse = (
    Any  # Use Any for now, as the exact response type depends on the LLM library
)


class ReasoningModule:
    """
    Combines hard prompts and soft prompts (derived from knowledge subgraphs)
    and interacts with a frozen LLM to infer the final answer to a user question.
    """

    def __init__(self, llm_model_name: str = "gemini-1.5-flash-latest"):
        """
        Initializes the ReasoningModule with a specific LLM.

        Args:
            llm_model_name (str): The name of the frozen LLM model to use for inference.
                                    Defaults to 'gemini-1.5-flash-latest'.
        """
        # Configure the generative AI library (requires API key setup externally)
        # genai.configure(api_key="YOUR_API_KEY") # API key should be handled securely, e.g., via environment variables

        # Initialize the frozen LLM model
        # Note: The model is assumed to be 'frozen' in the sense that its parameters
        # are not updated during this inference process.
        try:
            self.model = genai.GenerativeModel(llm_model_name)
            print(f"ReasoningModule initialized with LLM model: {llm_model_name}")
        except Exception as e:
            print(f"Error initializing LLM model {llm_model_name}: {e}")
            # Handle error appropriately, e.g., raise an exception or set model to None
            self.model = None  # Set model to None if initialization fails

    def _format_subgraph_for_prompt(self, subgraph: Subgraph) -> str:
        """
        Formats the knowledge subgraph into a string representation suitable for
        including in the LLM prompt as context.

        Args:
            subgraph (Subgraph): The knowledge subgraph (list of paths, list of triples).

        Returns:
            str: A formatted string representing the subgraph knowledge.
        """
        if not subgraph:
            return "No relevant knowledge found."

        formatted_knowledge = "Relevant Knowledge Graph Triples:\n"
        for i, path in enumerate(subgraph):
            formatted_knowledge += f"Path {i + 1}:\n"
            for triple in path:
                head, relation, tail = triple
                formatted_knowledge += f"- ({head}, {relation}, {tail})\n"
        return formatted_knowledge

    def infer(self, question: str, subgraph: Subgraph) -> str:
        """
        Infers the answer to a user question using hard prompts,
        knowledge derived from the subgraph (soft prompts), and a frozen LLM.

        Args:
            question (str): The user's question.
            subgraph (Subgraph): The knowledge subgraph retrieved by the adapter.

        Returns:
            str: The inferred answer from the LLM. Returns an error message if inference fails.
        """
        if self.model is None:
            return "Error: LLM model not initialized."

        # Define the hard prompt (system instructions and context)
        # This guides the LLM on how to use the provided knowledge and answer the question.
        hard_prompt = (
            "You are an expert reasoning system. Use the provided knowledge graph triples "
            "to answer the user's question. If the knowledge is insufficient, state that "
            "you cannot answer based on the provided information. Do not use external knowledge.\n\n"
        )

        # Format the subgraph knowledge to be included as part of the prompt (soft prompt context)
        knowledge_context = self._format_subgraph_for_prompt(subgraph)

        # Combine hard prompt, knowledge context, and user question
        # The knowledge context acts as the 'soft prompt' by providing specific,
        # relevant information derived from the knowledge graph embeddings.
        full_prompt = (
            f"{hard_prompt}{knowledge_context}\nUser Question: {question}\nAnswer:"
        )

        print(f"Sending prompt to LLM:\n{full_prompt}")

        try:
            # Perform inference using the frozen LLM
            # The generate_content method sends the prompt to the Gemini model
            response = self.model.generate_content(full_prompt)

            # Extract the answer from the LLM's response
            answer = response.text.strip()
            print(f"Received answer from LLM: {answer}")
            return answer

        except Exception as e:
            print(f"Error during LLM inference: {e}")
            return f"Error during inference: {e}"


# Example Usage (for testing the module in isolation)
if __name__ == "__main__":
    # Note: To run this example, you need to have the google-generativeai library installed
    # and your API key configured (e.g., via environment variable GOOGLE_API_KEY).

    # Create a dummy subgraph (list of paths with triples)
    dummy_subgraph_example: Subgraph = [
        [("entity1", "relationA", "entity2"), ("entity2", "relationB", "entity3")],
        [("entity1", "relationC", "entity3")],
    ]

    # Create a dummy question
    dummy_question = "What is the relationship between entity1 and entity3?"

    # Initialize the ReasoningModule
    # Use a suitable Gemini model name available via the API
    # Make sure to handle API key configuration before running
    try:
        # Assuming API key is configured externally (e.g., env var)
        reasoning_module = ReasoningModule(llm_model_name="gemini-1.5-flash-latest")

        # Perform inference
        inferred_answer = reasoning_module.infer(dummy_question, dummy_subgraph_example)

        print(f"\nQuestion: {dummy_question}")
        print(f"Inferred Answer: {inferred_answer}")

    except Exception as e:
        print(
            f"Could not run example: {e}. Ensure google-generativeai is installed and API key is configured."
        )

    # Example with empty subgraph
    print("\n--- Testing with empty subgraph ---")
    empty_subgraph: Subgraph = []
    dummy_question_2 = (
        "What is the capital of France?"  # Question not answerable by dummy knowledge
    )

    try:
        # Assuming API key is configured externally (e.g., env var)
        reasoning_module_2 = ReasoningModule(llm_model_name="gemini-1.5-flash-latest")
        inferred_answer_2 = reasoning_module_2.infer(dummy_question_2, empty_subgraph)
        print(f"\nQuestion: {dummy_question_2}")
        print(f"Inferred Answer: {inferred_answer_2}")
    except Exception as e:
        print(
            f"Could not run example with empty subgraph: {e}. Ensure google-generativeai is installed and API key is configured."
        )
