requirements.txt
networkx
transformers
google-generativeai
torch
python-dotenv
jupyter
pytest